+++ includes/crons/sales/gog_import.php
@@ -143,9 +143,14 @@ do {
 	}
 } while ($count > 0);
 
+$total_on_sale = count($on_sale);
+
 // remove any not found on sale
-$in  = str_repeat('?,', count($on_sale) - 1) . '?';
-$dbl->run("DELETE FROM `sales` WHERE `game_id` NOT IN ($in) AND `store_id` = 5", $on_sale);
+if (isset($total_on_sale) && $total_on_sale > 0)
+{
+	$in  = str_repeat('?,', count($on_sale) - 1) . '?';
+	$dbl->run("DELETE FROM `sales` WHERE `game_id` NOT IN ($in) AND `store_id` = 5", $on_sale);
+}
 
 echo "\n\n";
 echo "End of GOG import @ " . date('d-m-Y H:m:s') . ".\nHave a nice day.\n";
